test(python): split latest endpoint tests out of the historical file
test_when_fetching_historical_data.py covered both windowed and latest
endpoints, which are different shapes: one takes a from/to range, the
other a single asset id. Move the latest endpoints to their own file and
lift the shared fixtures and the 404-tolerant assertion helper into
conftest.py.

Also extends latest coverage while the endpoints are grouped: the
forecast endpoints, the facility endpoints, and a generous `since` per
endpoint that distinguishes a working parameter from one the API rejects
outright.
